@@ -1,173 +1,152 @@
-namespace LibraryManager.Api
+namespace LibraryManager.Api;
+
+public class Startup
 {
-    using LibraryManager.Api.Configurations;
-    using LibraryManager.Api.Services;
-    using LibraryManager.Domain.Abstractions.Services;
-    using LibraryManager.Infrastructure;
-    using LibraryManager.Infrastructure.Repositories;
-    using LibraryManager.Infrastructure.Repositories.Abstractions;
-    using LibraryManager.Infrastructure.Services;
-    using Microsoft.AspNetCore.Authentication.JwtBearer;
-    using Microsoft.AspNetCore.Builder;
-    using Microsoft.AspNetCore.Hosting;
-    using Microsoft.EntityFrameworkCore;
-    using Microsoft.Extensions.Configuration;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.IdentityModel.Tokens;
-    using Microsoft.OpenApi.Models;
-    using Newtonsoft.Json;
-    using System;
-    using System.Collections.Generic;
-    using System.Text;
-
-    public class Startup
+    public Startup(IConfiguration configuration)
     {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
+        Configuration = configuration;
+    }
 
-        public IConfiguration Configuration { get; }
+    public IConfiguration Configuration { get; }
 
-        public void ConfigureServices(IServiceCollection services)
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddCors();
+        services.AddControllers();
+        services.AddSwaggerGen(c =>
         {
-            services.AddCors();
-            services.AddControllers();
-            services.AddSwaggerGen(c =>
+            c.SwaggerDoc("v1", new OpenApiInfo { Title = "LibraryManager.Api", Version = "v1" });
+            c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
             {
-                c.SwaggerDoc("v1", new OpenApiInfo { Title = "LibraryManager.Api", Version = "v1" });
-                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-                {
-                    Description =
-        "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
-                    Name = "Authorization",
-                    In = ParameterLocation.Header,
-                    Type = SecuritySchemeType.ApiKey,
-                    Scheme = "Bearer"
-                });
+                Description =
+    "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
+                Name = "Authorization",
+                In = ParameterLocation.Header,
+                Type = SecuritySchemeType.ApiKey,
+                Scheme = "Bearer"
+            });
 
-                c.AddSecurityRequirement(new OpenApiSecurityRequirement()
+            c.AddSecurityRequirement(new OpenApiSecurityRequirement()
+            {
                 {
+                    new OpenApiSecurityScheme
                     {
-                        new OpenApiSecurityScheme
+                        Reference = new OpenApiReference
                         {
-                            Reference = new OpenApiReference
-                            {
-                                Type = ReferenceType.SecurityScheme,
-                                Id = "Bearer"
-                            },
-                            Scheme = "oauth2",
-                            Name = "Bearer",
-                            In = ParameterLocation.Header,
-
+                            Type = ReferenceType.SecurityScheme,
+                            Id = "Bearer"
                         },
-                        new List<string>()
-                    }
-                });
+                        Scheme = "oauth2",
+                        Name = "Bearer",
+                        In = ParameterLocation.Header,
+
+                    },
+                    new List<string>()
+                }
             });
+        });
 
-            services.AddControllersWithViews()
-                .AddNewtonsoftJson(options =>
-                    options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore
-            );
+        services.AddControllersWithViews()
+            .AddNewtonsoftJson(options =>
+                options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore
+        );
 
-            services.AddDbContext<LibraryManagerDbContext>(
-                options =>
-                    options.UseNpgsql(
-                        GetConnectionString(),
-                        x => x.MigrationsAssembly("LibraryManager.Infrastructure")));
-
-
-            services.Configure<Settings>(Configuration.GetSection("Settings"));
+        services.AddDbContext<LibraryManagerDbContext>(
+            options =>
+                options.UseNpgsql(
+                    GetConnectionString(),
+                    x => x.MigrationsAssembly("LibraryManager.Infrastructure")));
 
 
-            var settings = Environment.GetEnvironmentVariable("Settings")
-                            ?? Configuration.GetSection("Settings").GetSection("Secret").Value;
+        services.Configure<Settings>(Configuration.GetSection("Settings"));
 
-            services
-                .AddAuthentication(x =>
-                {
-                    x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                    x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(x =>
-                {
-                    x.RequireHttpsMetadata = false;
-                    x.SaveToken = true;
-                    x.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidateIssuerSigningKey = true,
-                        IssuerSigningKey = new SymmetricSecurityKey(
-                            Encoding.ASCII.GetBytes(settings)),
-                        ValidateIssuer = false,
-                        ValidateAudience = false
-                    };
-                });
 
-            //services.AddDistributedRedisCache(options =>
-            //{
-            //    options.Configuration =
-            //        Configuration.GetConnectionString("LibraryManagerRedis");
-            //    options.InstanceName = "LibraryManagerRedis";
-            //});
+        var settings = Environment.GetEnvironmentVariable("Settings")
+                        ?? Configuration.GetSection("Settings").GetSection("Secret").Value;
 
-            // Repositories
-            services.AddTransient(typeof(IBaseRepository<>), typeof(BaseRepository<>));
-            services.AddTransient<ITransactionRepository, TransactionRepository>();
-            services.AddTransient<IBookRepository, BookRepository>();
-            services.AddTransient<IStudentRepository, StudentRepository>();
-            services.AddTransient<ITeacherRepository, TeacherRepository>();
-            services.AddTransient<IAuthorRepository, AuthorRepository>();
-
-            // Services
-            services.AddTransient<ITransactionService, TransactionService>();
-            services.AddTransient<IBookService, BookService>();
-            services.AddTransient<IStudentService, StudentService>();
-            services.AddTransient<ITeacherService, TeacherService>();
-            services.AddTransient<IAuthorService, AuthorService>();
-            services.AddTransient<ITokenService, TokenService>();
-            services.AddTransient<IAuthService, AuthService>();
-            services.AddTransient<IEncryptService, EncryptService>();
-
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-
-            app.UseDeveloperExceptionPage();
-            app.UseSwagger();
-            app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "LibraryManager.Api v1"));
-
-            app.UseHttpsRedirection();
-
-            app.UseRouting();
-
-            app.UseAuthentication();
-
-            app.UseAuthorization();
-
-            app.UseCors(
-                options => options.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader()
-            );
-
-            app.UseEndpoints(endpoints =>
+        services
+            .AddAuthentication(x =>
             {
-                endpoints.MapControllers().RequireAuthorization();
+                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+            })
+            .AddJwtBearer(x =>
+            {
+                x.RequireHttpsMetadata = false;
+                x.SaveToken = true;
+                x.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuerSigningKey = true,
+                    IssuerSigningKey = new SymmetricSecurityKey(
+                        Encoding.ASCII.GetBytes(settings)),
+                    ValidateIssuer = false,
+                    ValidateAudience = false
+                };
             });
-        }
 
-        private string GetConnectionString()
+        //services.AddDistributedRedisCache(options =>
+        //{
+        //    options.Configuration =
+        //        Configuration.GetConnectionString("LibraryManagerRedis");
+        //    options.InstanceName = "LibraryManagerRedis";
+        //});
+
+        // Repositories
+        services.AddTransient(typeof(IBaseRepository<>), typeof(BaseRepository<>));
+        services.AddTransient<ITransactionRepository, TransactionRepository>();
+        services.AddTransient<IBookRepository, BookRepository>();
+        services.AddTransient<IStudentRepository, StudentRepository>();
+        services.AddTransient<ITeacherRepository, TeacherRepository>();
+        services.AddTransient<IAuthorRepository, AuthorRepository>();
+
+        // Services
+        services.AddTransient<ITransactionService, TransactionService>();
+        services.AddTransient<IBookService, BookService>();
+        services.AddTransient<IStudentService, StudentService>();
+        services.AddTransient<ITeacherService, TeacherService>();
+        services.AddTransient<IAuthorService, AuthorService>();
+        services.AddTransient<ITokenService, TokenService>();
+        services.AddTransient<IAuthService, AuthService>();
+        services.AddTransient<IEncryptService, EncryptService>();
+
+    }
+
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+
+        app.UseDeveloperExceptionPage();
+        app.UseSwagger();
+        app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "LibraryManager.Api v1"));
+
+        app.UseHttpsRedirection();
+
+        app.UseRouting();
+
+        app.UseAuthentication();
+
+        app.UseAuthorization();
+
+        app.UseCors(
+            options => options.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader()
+        );
+
+        app.UseEndpoints(endpoints =>
         {
-            var uriString = Environment.GetEnvironmentVariable("DATABASE_URL")
-                                ?? Configuration.GetConnectionString("DATABASE_URL");
-            var uri = new Uri(uriString);
-            var db = uri.AbsolutePath.Trim('/');
-            var user = uri.UserInfo.Split(':')[0];
-            var passwd = uri.UserInfo.Split(':')[1];
-            var port = uri.Port > 0 ? uri.Port : 5432;
-            var connStr = string.Format("Server={0};Database={1};User Id={2};Password={3};Port={4};SSL Mode=Require;Trust Server Certificate=True;",
-                uri.Host, db, user, passwd, port);
-            return connStr;
-        }
+            endpoints.MapControllers().RequireAuthorization();
+        });
+    }
+
+    private string GetConnectionString()
+    {
+        var uriString = Environment.GetEnvironmentVariable("DATABASE_URL")
+                            ?? Configuration.GetConnectionString("DATABASE_URL");
+        var uri = new Uri(uriString);
+        var db = uri.AbsolutePath.Trim('/');
+        var user = uri.UserInfo.Split(':')[0];
+        var passwd = uri.UserInfo.Split(':')[1];
+        var port = uri.Port > 0 ? uri.Port : 5432;
+        var connStr = string.Format("Server={0};Database={1};User Id={2};Password={3};Port={4};SSL Mode=Require;Trust Server Certificate=True;",
+            uri.Host, db, user, passwd, port);
+        return connStr;
     }
 }
