@@ -1,20 +1,18 @@
-﻿namespace LibraryManager.Domain.Entities
+﻿namespace LibraryManager.Domain.Entities;
+
+public sealed class Book : BaseEntity
 {
-    using System.Collections.Generic;
+    public int Reference { get; set; }
 
-    public sealed class Book : BaseEntity
-    {
-        public int Reference { get; set; }
-        public Author Author { get; set; }
+    public Author Author { get; set; }
 
-        public long AuthorId { get; set; }
+    public long AuthorId { get; set; }
 
-        public int CheckoutPeriod { get; set; }
+    public int CheckoutPeriod { get; set; }
 
-        public string Description { get; set; }
+    public string Description { get; set; }
 
-        public string Title { get; set; }
+    public string Title { get; set; }
 
-        public IList<Transaction> Transactions { get; set; }
-    }
+    public IList<Transaction> Transactions { get; set; }
 }
