@@ -1,31 +1,27 @@
-﻿namespace LibraryManager.Infrastructure
+﻿namespace LibraryManager.Infrastructure;
+
+public class LibraryManagerDbContext : DbContext
 {
-    using LibraryManager.Domain.Entities;
-    using Microsoft.EntityFrameworkCore;
-
-    public class LibraryManagerDbContext : DbContext
+    public LibraryManagerDbContext(DbContextOptions dbContextOptions) : base(dbContextOptions)
     {
-        public LibraryManagerDbContext(DbContextOptions dbContextOptions) : base(dbContextOptions)
-        {
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder);
-
-            modelBuilder.ApplyConfigurationsFromAssembly(typeof(BaseEntity).Assembly);
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            base.OnConfiguring(optionsBuilder);
-        }
-
-        public DbSet<Author> Authors { get; set; }
-        public DbSet<Book> Books { get; set; }
-        public DbSet<Student> Students { get; set; }
-        public DbSet<Teacher> Teachers { get; set; }
-        public DbSet<Transaction> Transactions { get; set; }
-        
     }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        base.OnModelCreating(modelBuilder);
+
+        modelBuilder.ApplyConfigurationsFromAssembly(typeof(BaseEntity).Assembly);
+    }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        base.OnConfiguring(optionsBuilder);
+    }
+
+    public DbSet<Author> Authors { get; set; }
+    public DbSet<Book> Books { get; set; }
+    public DbSet<Student> Students { get; set; }
+    public DbSet<Teacher> Teachers { get; set; }
+    public DbSet<Transaction> Transactions { get; set; }
+
 }
