@@ -1,13 +1,9 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿namespace LibraryManager.Infrastructure.Repositories.Abstractions;
 
-namespace LibraryManager.Infrastructure.Repositories.Abstractions
+public interface IBaseRepository<T> where T : class
 {
-    public interface IBaseRepository<T> where T : class
-    {
-        Task<IEnumerable<T>> GetAll();
-        void Insert(T obj);
-        void Update(T obj);
-        Task<bool> Save();
-    }
+    Task<IEnumerable<T>> GetAll();
+    void Insert(T obj);
+    void Update(T obj);
+    Task<bool> Save();
 }
