@@ -1,38 +1,50 @@
-﻿using LibraryManager.Domain.Abstractions.Services;
-using LibraryManager.Domain.Dtos;
-using LibraryManager.Domain.Entities;
-using LibraryManager.Infrastructure.Repositories.Abstractions;
-using System.Threading.Tasks;
+﻿namespace LibraryManager.Infrastructure.Services;
 
-namespace LibraryManager.Infrastructure.Services
+public class AuthService : IAuthService
 {
-    public class AuthService : IAuthService
+    private readonly ITeacherRepository repository;
+    private readonly IEncryptService service;
+
+    public AuthService(ITeacherRepository repository, IEncryptService service)
     {
-        private readonly ITeacherRepository repository;
-        private readonly IEncryptService service;
+        this.repository = repository;
+        this.service = service;
+    }
 
-        public AuthService(ITeacherRepository repository, IEncryptService service)
+    public async Task<bool> Register(RegisterDto registerDto)
+    {
+        var reference = await this.repository.GetLastReference() + 1;
+
+        var user = new Teacher
         {
-            this.repository = repository;
-            this.service = service;
+            Email = registerDto.Email,
+            Name = registerDto.Name,
+            Password = registerDto.Password,
+            Role = registerDto.Role,
+            Reference = reference
+        };
+
+        try
+        {
+            this.repository.Insert(user);
+            return await this.repository.Save();
         }
-
-        public async Task<bool> Register(RegisterDto registerDto)
+        catch
         {
-            var reference = await this.repository.GetLastReference() + 1;
+            throw;
+        }
+    }
 
-            var user = new Teacher
-            {
-                Email = registerDto.Email,
-                Name = registerDto.Name,
-                Password = registerDto.Password,
-                Role = registerDto.Role,
-                Reference = reference
-            };
+    public async Task<bool> ChangePassword(ChangePasswordDto changePasswordDto)
+    {
+        var oldEncryptedPassword = this.service.Encrypt(changePasswordDto.OldPassword);
+        var user = await this.repository.GetByEmailAndPassword(changePasswordDto.Email, oldEncryptedPassword);
 
+        if (user != null)
+        {
             try
             {
-                this.repository.Insert(user);
+                user.Password = this.service.Encrypt(changePasswordDto.NewPassword);
                 return await this.repository.Save();
             }
             catch
@@ -41,25 +53,6 @@ namespace LibraryManager.Infrastructure.Services
             }
         }
 
-        public async Task<bool> ChangePassword(ChangePasswordDto changePasswordDto)
-        {
-            var oldEncryptedPassword = this.service.Encrypt(changePasswordDto.OldPassword);
-            var user = await this.repository.GetByEmailAndPassword(changePasswordDto.Email, oldEncryptedPassword);
-
-            if(user != null)
-            {
-                try
-                {
-                    user.Password = this.service.Encrypt(changePasswordDto.NewPassword);
-                    return await this.repository.Save();
-                }
-                catch
-                {
-                    throw;
-                }
-            }
-
-            return false;
-        }
+        return false;
     }
 }
