@@ -1,47 +1,73 @@
-﻿using LibraryManager.Domain.Abstractions.Services;
-using LibraryManager.Domain.Dtos.Books;
-using LibraryManager.Domain.Dtos.Transactions;
-using LibraryManager.Domain.Entities;
-using LibraryManager.Infrastructure.Repositories.Abstractions;
-using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿namespace LibraryManager.Infrastructure.Services;
 
-namespace LibraryManager.Infrastructure.Services
+public class TransactionService : ITransactionService
 {
-    public class TransactionService : ITransactionService
+    private readonly ITransactionRepository repository;
+
+    public TransactionService(ITransactionRepository repository)
     {
-        private readonly ITransactionRepository repository;
+        this.repository = repository;
+    }
 
-        public TransactionService(ITransactionRepository repository)
-        {
-            this.repository = repository;
-        }
-
-        public async Task<IEnumerable<Transaction>> GetAll()
+    public async Task<IEnumerable<Transaction>> GetAll()
+    {
+        try
         {
             return await this.repository.GetAll();
         }
+        catch
+        {
+            throw new Exception("An error occurred while all transactions.");
+        }
+    }
 
-        public async Task<IEnumerable<Transaction>> GetAllActiveTransactions()
+    public async Task<IEnumerable<Transaction>> GetAllActiveTransactions()
+    {
+        try
         {
             return await this.repository.GetAllActiveTransactions();
         }
+        catch
+        {
+            throw new Exception("An error occurred while getting all active transactions.");
+        }
+    }
 
 
-        public async Task<IEnumerable<Transaction>> GetAllByBook(long bookId)
+    public async Task<IEnumerable<Transaction>> GetAllByBook(long bookId)
+    {
+        try
         {
             var transactions = await this.repository.GetAllByBook(bookId);
             return transactions;
         }
+        catch
+        {
+            throw new Exception("An error occurred while getting all transactions by book.");
+        }
+    }
 
-        public async Task<IEnumerable<LateBookWithStudentNameDto>> GetLateBooksWithStudentName(long teacherId)
+    public async Task<IEnumerable<LateBookWithStudentNameDto>> GetLateBooksWithStudentName(long teacherId)
+    {
+        try
         {
             return await this.repository.GetLateBooksWithStudentName(teacherId);
         }
+        catch
+        {
+            throw new Exception("An error occurred while getting the late book list.");
+        }
+    }
 
-        public async Task<IEnumerable<GetTransactionDto>> GetTransactionsWithDetailsByStudent(long studentId)
+    public async Task<IEnumerable<GetTransactionDto>> GetTransactionsWithDetailsByStudent(long studentId)
+    {
+        try
         {
             return await this.repository.GetTransactionsWithDetailsByStudent(studentId);
+        }
+        catch
+        {
+            throw new Exception("An error occurred while getting the transaction list.");
         }
     }
 }
